+++ src/cmdtools/__init__.py
@@ -1,3 +1,3 @@
 from .analysis import pcca, schur, optimization
-from .estimation import diffusionmaps, galerkin, gillespie, newton_generator, picking_algorithm, sqra, voronoi
+from .estimation import diffusionmaps, galerkin, gillespie, newton_generator, picking_algorithm, sqra, voronoi, ajc
 from .systems import diffusion