+++ src/cmdtools/estimation/sqra.py
@@ -23,7 +23,8 @@ def sqra(u, A, beta, phi):
         Q = Q.tocsc()
     return Q
 
-def adjacency_nd(dims, torus = False):
+
+def adjacency_nd(dims, torus=False):
     nd = len(dims)
     N = np.prod(dims)
 
@@ -37,28 +38,41 @@ def adjacency_nd(dims, torus = False):
     cut = np.zeros(N*2*nd, dtype=bool)
 
     for i in range(N):
-        multi = np.unravel_index(i, dims) # find multiindex of current cell
-        mn = multi + neighbours # add neighbour offset
+        multi = np.unravel_index(i, dims)  # find multiindex of current cell
+        mn = multi + neighbours  # add neighbour offset
         if not torus:
-            cut[i*2*nd:(i+1)*2*nd] = np.sum((mn < 0) + ( mn >= dims), axis=1) # check if boundary is hit
-        mn = np.mod(multi + neighbours, dims) # glue together boundary
-        neighbour_flat = np.ravel_multi_index(mn.T, dims) # back to flat inds
-        #print(neighbour_flat)
+            cut[i*2*nd:(i+1)*2*nd] = np.sum((mn < 0) + (mn >= dims), axis=1)  # check if boundary is hit
+        mn = np.mod(multi + neighbours, dims)  # glue together boundary
+        neighbour_flat = np.ravel_multi_index(mn.T, dims)  # back to flat inds
+        # print(neighbour_flat)
         row[i*2*nd:(i+1)*2*nd] = i
         col[i*2*nd:(i+1)*2*nd] = neighbour_flat
 
-    if not torus: # cut out the points which were glued at boundaries
+    if not torus:  # cut out the points which were glued at boundaries
         sel = ~cut
         data = data[sel]
         row = row[sel]
         col = col[sel]
 
-    #return row, col
-    A = sp.coo_matrix((data, (row, col))).tocsr()
-    A[np.diag_indices_from(A)] = False  # todo: this throws an inefficiency warning.
-    return A
+    A = sp.coo_matrix((data, (row, col)))
+    A.setdiag(0)
+    return A.tocsr()
+
 
 class SQRA:
+    """ approximate the generator/rate-matrix Q of the Overdamped-Langevin dynamics
+    Args:
+        u (ndarray): The potential function evaluated at the grid points.
+        beta (float): The inverse temperatur of the system (scales the rates nonlinear).
+        phi (float): Linear scaling factor of the rates (depending on the grid volume).
+        A (matrix, optional): Adjacency matrix of the grid. If left empty, automatically compute it based on the shape of `u`.
+        torus(list of bool): Whether to glue the corresponding dimensions together at their resp. boundaries. Only used in the automatic generation of `A`.
+
+    Attributes:
+        A: Adjacency matrix
+        Q: The computed generator
+    """
+
     def __init__(self, u, beta=1, phi=1, A=None, torus=False):
         self.u = u
         self.beta = beta
@@ -72,6 +86,7 @@ class SQRA:
 
     def sqra(self):
         return sqra(self.u.flatten(), self.A, self.beta, self.phi)
+
 
 # conversion between coldness (beta) and epsilon in sde formulation
 # necessary for the computation of phi